pyproject.toml
[project]
name = "osync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Akshay Kumar", email = "132216831+bmblb3@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[project.scripts]
osync = "osync:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
